fix(editor): send displayed value + UOM, controller scales device-side (#165)
## Problem

The `/cmd` wire contract is **send the displayed value, tail the range
UOM, the controller scales device-side** — proven on live hardware:

| URL | Result |
|---|---|
| `/cmd/DON/100/100` | 39 % (100 read as a UOM-100 0-255 byte) |
| `/cmd/DON/100/51` | 100 % (100 read as UOM-51 percent) |

and the eisy web UI sends `/cmd/setTemp/10.4/17` (the displayed float,
not a pre-scaled int).

`Editor._encode_in_range` instead rewrote the number client-side:
`round(v * 10**prec)`, the half-degree `* 2`, and an unconditional
`round()`. So a PG3 `virtualtemp` `setTemp` (UOM 17, prec 1) sent `10`
for input `1` ("set 1, reads back 10") and could never emit a fractional
`10.4`.

## Fix

`encode` now **validates only** — enum-name resolution, `min`/`max`,
`subset`, all in *displayed* units — and returns the value **as-is**:
`int` when integral (URL stays `/72`, not `/72.0`), else `float` so
`/cmd/setTemp/10.4/17` survives. `encode` / `encode_param` /
`_encode_in_range` and the send plumbing (`_commands._encode`,
`node.send_command` wire args, `client.send_node_command`) widen to `int
| float`.

This **reverts the 2026-05-09 prec-symmetric encode change** — that
double-scaled against the controller's own UOM scaling (including for
native Insteon, where `/cmd/.../<displayed>/<uom>` is also the
contract). `decode` is **unchanged**: it stays a precision-aware
*display* helper and is now intentionally decoupled from `encode` (the
property read path normalises by the wire UOM and never calls `decode`).

## Verification

- Full suite **821 passed**, ruff + mypy clean. The 11
`test_editor_codec` tests that asserted the old client-scaling contract
are rewritten to the new one (encode validates + sends as-is; decode
independent).
- Live (eisy `192.168.30.174`): `setTemp`
`encode_param(10.4)=(10.4,'17')`, `(1)=(1,'17')`, `(70.5)=(70.5,'17')`;
native dimmer `DON` `encode_param(100)=(100,'51')`, `(75)=(75,'51')`.

Surfaced during hacs#67 live testing. Independent of #163 / #164; all
three target `6.0.0b6`.

⚠️ No Insteon-thermostat hardware here to re-verify the half-degree
UOM-101 path; the change applies the same proven `/cmd displayed-value +
UOM` rule (controller does the `*2`).

+++ pyisyox/client.py
@@ -629,7 +629,7 @@ class IoXClient:
             _LOGGER.debug("optional endpoint %s unavailable: %s", path, exc)
             return ""
 
-    async def send_node_command(self, address: str, command_id: str, *params: int | str) -> str:
+    async def send_node_command(self, address: str, command_id: str, *params: float | str) -> str:
         """Issue ``GET /rest/nodes/{addr}/cmd/{cmd}[/{p1}[/{p2}...]]``.
 
         Params are stringified and joined as-is — the editor codec runs
